Extract operator lifecycle boundary from cmd/gestaltd
