+++ cmd/gestaltd/localconfig.go
@@ -1,51 +1,11 @@
 package main
 
-import (
-	"crypto/rand"
-	"encoding/hex"
-	"fmt"
-	"log"
-	"os"
-	"path/filepath"
-)
+import "github.com/valon-technologies/gestalt/internal/operator"
 
 func defaultLocalConfigPath() string {
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return ""
-	}
-	return filepath.Join(home, ".gestalt", "config.yaml")
+	return operator.DefaultLocalConfigPath()
 }
 
 func generateDefaultConfig(configDir string) (string, error) {
-	configPath := filepath.Join(configDir, "config.yaml")
-	if err := os.MkdirAll(configDir, 0755); err != nil {
-		return "", fmt.Errorf("create config dir: %w", err)
-	}
-
-	key := make([]byte, 32)
-	if _, err := rand.Read(key); err != nil {
-		return "", fmt.Errorf("generate encryption key: %w", err)
-	}
-
-	dbPath := filepath.Join(configDir, "gestalt.db")
-	cfg := fmt.Sprintf(`auth:
-  provider: local
-datastore:
-  provider: sqlite
-  config:
-    path: %q
-secrets:
-  provider: env
-server:
-  port: 8080
-  dev_mode: true
-  encryption_key: %q
-`, dbPath, hex.EncodeToString(key))
-
-	if err := os.WriteFile(configPath, []byte(cfg), 0600); err != nil {
-		return "", fmt.Errorf("write config: %w", err)
-	}
-	log.Printf("generated default config at %s", configPath)
-	return configPath, nil
+	return operator.GenerateDefaultConfig(configDir)
 }
