+++ internal/operator/localconfig.go
@@ -1,0 +1,51 @@
+package operator
+
+import (
+	"crypto/rand"
+	"encoding/hex"
+	"fmt"
+	"log"
+	"os"
+	"path/filepath"
+)
+
+func DefaultLocalConfigPath() string {
+	home, err := os.UserHomeDir()
+	if err != nil {
+		return ""
+	}
+	return filepath.Join(home, ".gestalt", "config.yaml")
+}
+
+func GenerateDefaultConfig(configDir string) (string, error) {
+	configPath := filepath.Join(configDir, "config.yaml")
+	if err := os.MkdirAll(configDir, 0o755); err != nil {
+		return "", fmt.Errorf("create config dir: %w", err)
+	}
+
+	key := make([]byte, 32)
+	if _, err := rand.Read(key); err != nil {
+		return "", fmt.Errorf("generate encryption key: %w", err)
+	}
+
+	dbPath := filepath.Join(configDir, "gestalt.db")
+	cfg := fmt.Sprintf(`auth:
+  provider: local
+datastore:
+  provider: sqlite
+  config:
+    path: %q
+secrets:
+  provider: env
+server:
+  port: 8080
+  dev_mode: true
+  encryption_key: %q
+`, dbPath, hex.EncodeToString(key))
+
+	if err := os.WriteFile(configPath, []byte(cfg), 0o600); err != nil {
+		return "", fmt.Errorf("write config: %w", err)
+	}
+	log.Printf("generated default config at %s", configPath)
+	return configPath, nil
+}
